footybitez/requirements.txt
moviepy==1.0.3
edge-tts==6.1.9
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
requests==2.31.0
pillow==10.2.0
python-dotenv==1.0.1
gTTS
groq
duckduckgo_search
beautifulsoup4
wikipedia
